+++ packages/schema-types/CHANGELOG.md
@@ -1,4 +1,10 @@
-# @zotero-suite/schema-types
+# zotero-schema-types
+
+## 1.0.2
+
+### Patch Changes
+
+- Changed scoping of packages
 
 ## 1.0.1